test(browser): split container tests by module
Move the browser container unit tests into a child test module file so container.rs stays under the repository file-size limit while preserving private test access.
